requirements.txt
uvicorn==0.17.6
fastapi==0.78.0
httpx==0.23.0
PyYAML==6.0
streamlit==1.10.0 # for streamlit usage. Not required for rest api.